requirements.txt
snowflake.core==1.0.2
snowflake-snowpark-python==1.26.0